+++ samples/Slice.Sample/Program.cs
@@ -1,14 +1,13 @@
-using Slice.Generated;
 using Slice.Sample.Services;
 
 var builder = WebApplication.CreateSlimBuilder(args);
 
-builder.Services.AddSliceGenerated();
+builder.Services.AddSlice();
 builder.Services.AddSingleton(TimeProvider.System);
 builder.Services.AddSingleton<IUserStore, InMemoryUserStore>();
 
 var app = builder.Build();
 
-app.MapSlicesGenerated(); // <-- ここで全Featureが自動登録される
+app.MapSlices(); // <-- ここで全Featureが自動登録される
 
 app.Run();